% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawyers.adjacency.advice.R
\docType{data}
\name{lawyers.adjacency.advice}
\alias{lawyers.adjacency.advice}
\title{Adjacency matrix detailing the presence or absence of advice links
between the `Lazega Lawyers'.}
\format{A 71 x 71 binary matrix, with 0 down the diagonal.}
\source{
E. Lazega, The Collegial Phenomenon: The Social Mechanisms of Cooperation
 Among Peers in a Corporate Law Partnership, Oxford University Press, Oxford, England, 2001.
}
\usage{
lawyers.adjacency.advice
}
\description{
Data on whether or not 71 lawyers in a northeastern American law firm
asked each other for advice.
}
\keyword{datasets}
