% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.twitter.adjacency.R
\docType{data}
\name{us.twitter.adjacency}
\alias{us.twitter.adjacency}
\title{Directed adjacency matrix detailing the presence or absence
of Twitter friend/follower links between US politicians.}
\format{A 69 x 69 binary matrix, with 0 down the diagonal.}
\source{
With thanks to Dr. Derek Greene. School of Computer Science,
University College Dublin.
}
\usage{
us.twitter.adjacency
}
\description{
Network data on whether or not 69 US politicians are
friends/followers on Twitter.
}
\keyword{datasets}
