% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calclambda.R
\name{calclambda}
\alias{calclambda}
\title{Title Compute mixing proportions}
\usage{
calclambda(tau, x.mix)
}
\arguments{
\item{tau}{A matrix of logistic regression coefficients, with G rows and number of columns equal to the number of covariates in the mixing proportions model plus 1, for the intercept.}

\item{x.mix}{A matrix of covariates in the mixing proportions model (including dummy variables for any factor covariates), with a column of 1's appended at the front.}
}
\value{
An n x G matrix of mixing proportions.
}
\description{
Function to compute the each observation's mixing proportions which are modeled as a logistic function of their covariates.
}
\references{
Isobel Claire Gormley and Thomas Brendan Murphy. (2010) A Mixture of Experts Latent Position Cluster Model for Social Network Data. Statistical Methodology, 7 (3), pp.385-405.
}
\seealso{
\code{\link{MEclustnet}}
}
